Implement an application using Binary Search Technique
Write a program that inputs a sorted array of integers. Ask user to input a value to search. Use Binary search to search for the value. If value is present display the position where it is present in the array.
 
(For output: Create an array of 10 elements. Show cases where value is present and also when value is not present)